move open logic to SharedWal
